Extract weight audit, batch-invariant linear, and clean dead paths
